put each content fetcher into its own module
