+++ container/shim/src/fetchers/lassie.js
@@ -1,0 +1,3 @@
+export function respondFromLassie(req, res, { cid, format }) {
+  // todo
+}
